[feat] 모달을 닫았을 때 카드를 모두 닫고 셔플 구현
