+++ week3/src/pages/CardGamePage.jsx
@@ -1,23 +1,19 @@
-import { useState, useMemo, useEffect } from "react";
+/* eslint-disable react/prop-types */
+import { useState, useEffect } from "react";
 import styled from "styled-components";
-import { IMAGEDATA } from "../contants/images";
 
-// eslint-disable-next-line react/prop-types
-function MainPage({ score, setScore }) {
-  const getMixedCardList = () => {
-    return [...IMAGEDATA, ...IMAGEDATA]
-      .map((card) => ({ ...card, status: false }))
-      .sort(() => 0.5 - Math.random());
-  };
-
-  const [cards, setCards] = useState(useMemo(getMixedCardList, []));
+function MainPage({ cards, setCards, score, setScore }) {
   const [clicked, setClicked] = useState([]);
-  const [modalVisible, setModalVisible] = useState(false);
 
   useEffect(() => {
     if (clicked.length === 2) {
       const [first, second] = clicked;
-      if (cards[first].id !== cards[second].id) {
+      // 카드 짝이 맞는지 확인
+      if (cards[first].id === cards[second].id) {
+        // 카드 짝이 맞으면 점수 1추가
+        setScore(score + 1);
+      } else {
+        // 카드가 매치되지 않는 경우 카드를 다시 뒤집기
         const newCards = cards.map((card, index) => {
           if (index === first || index === second) {
             return { ...card, status: false };
@@ -25,17 +21,10 @@ function MainPage({ score, setScore }) {
           return card;
         });
         setTimeout(() => setCards(newCards), 1000);
-      } else {
-        if (score < 5) {
-          setScore(score + 1);
-          if (score == 5) {
-            toggleModal();
-          }
-        }
       }
-      setClicked([]);
+      setClicked([]); // 클릭된 카드 초기화
     }
-  }, [clicked, cards]);
+  }, [clicked, cards, setScore]);
 
   const handleClick = (index) => {
     if (!cards[index].status && clicked.length < 2) {
@@ -45,10 +34,6 @@ function MainPage({ score, setScore }) {
       setCards(newCards);
       setClicked([...clicked, index]);
     }
-  };
-
-  const toggleModal = () => {
-    setModalVisible(!modalVisible);
   };
 
   return (
@@ -82,14 +67,6 @@ function MainPage({ score, setScore }) {
           ))}
         </CardWrapper>
       </Wrapper>
-      {modalVisible && (
-        <ModalWrapper>
-          <ModalDiv>
-            <Title>축하합니다</Title>
-            <RegameBtn onClick={toggleModal}>게임으로 돌아가기</RegameBtn>
-          </ModalDiv>
-        </ModalWrapper>
-      )}
     </>
   );
 }
@@ -100,47 +77,6 @@ const Wrapper = styled.div`
   justify-content: center;
   align-items: center;
   gap: 100px;
-`;
-// 모달
-
-const Title = styled.p`
-  color: ${(props) => props.theme.colors.white};
-  font-size: ${({ theme }) => theme.fonts.lg};
-`;
-
-const ModalWrapper = styled.div`
-  /* display: flex;
-  justify-content: center;
-  align-items: center; */
-  /* height: 100%;
-  width: 100%;
-  position: fixed;
-  background-color: ${({ theme }) => theme.colors.black}; */
-`;
-
-const ModalDiv = styled.div`
-  position: fixed;
-  top: 300px;
-  left: 550px;
-  display: flex;
-  z-index: 1;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 200px;
-  width: 350px;
-  background-color: ${({ theme }) => theme.colors.skyBlue};
-`;
-const RegameBtn = styled.button`
-  width: 10vw;
-  height: 4vh;
-  margin-top: 5vh;
-  background-color: ${({ theme }) => theme.colors.lightYellow};
-  border: none;
-  border-radius: 30px;
-  &:hover {
-    background-color: ${({ theme }) => theme.colors.skyBlue};
-  }
 `;
 
 // button
